Separating out-of-date message code into its own file
