+++ readthedocs/core/static-src/core/js/doc-embed/version-compare.js
@@ -1,0 +1,42 @@
+var rtd = require('./rtd-data');
+
+
+function init() {
+
+    /// Out of date message
+
+      var versionURL = [rtd.api_host + "/api/v1/version/", rtd['project'],
+                        "/highest/", rtd['version'], "/?callback=?"].join("");
+
+      $.getJSON(versionURL, onData);
+
+      function onData (data) {
+        if (data.is_highest) {
+          return;
+        }
+
+        var currentURL = window.location.pathname.replace(rtd['version'], data.slug),
+            warning = $('<div class="admonition warning"> <p class="first \
+                         admonition-title">Note</p> <p class="last"> \
+                         You are not using the most up to date version \
+                         of the library. <a href="#"></a> is the newest version.</p>\
+                         </div>');
+
+        warning
+          .find('a')
+          .attr('href', currentURL)
+          .text(data.version);
+
+        body = $("div.body");
+        if (!body.length) {
+          body = $("div.document");
+        }
+        body.prepend(warning);
+      }
+
+}
+
+
+module.exports = {
+    init: init
+};
